rearchitecting the native IEEE-754 constexpr/nonconstexpr setup
