+++ include/universal/internal/f2s/f2s.hpp
@@ -306,8 +306,8 @@ namespace sw {
 			template<typename Real>
 			f2s& convert_ieee754(Real v) noexcept {
 				bool sign{ false };
-				uint64_t biased{ 0 }, f64{ 0 };
-				extractFields(v, sign, biased, f64);
+				uint64_t biased{ 0 }, f64{ 0 }, bits{ 0 };
+				extractFields(v, sign, biased, f64, bits);
 				s_ = sign;
 				e_ = static_cast<int>(biased) - ieee754_parameter<Real>::bias;
 				f_ = static_cast<UnsignedInt>(ieee754_parameter<Real>::hmask | f64); // add the hidden bit
