+++ include/universal/internal/value/value.hpp
@@ -13,7 +13,9 @@
 
 #include <universal/common/exceptions.hpp>
 #include <universal/number/support/decimal.hpp>
+#include <universal/utility/find_msb.hpp>
 #include <universal/native/ieee754.hpp>
+#include <universal/native/nonconstexpr/extract_fp_components.hpp>
 #include <universal/internal/bitblock/bitblock.hpp>
 
 namespace sw { namespace universal { namespace internal {
@@ -130,7 +132,7 @@ public:
 		_sign = (0x8000000000000000 & rhs);  // 1 is negative, 0 is positive
 		if (_sign) {
 			// process negative number: process 2's complement of the input
-			_scale = int(sw::universal::findMostSignificantBit(-rhs)) - 1;
+			_scale = int(sw::universal::find_msb(-rhs)) - 1;
 			uint64_t _fraction_without_hidden_bit = uint64_t(_scale == 0 ? 0 : (-rhs << (64 - _scale)));
 			_fraction = copy_integer_fraction<fbits>(_fraction_without_hidden_bit);
 			//take_2s_complement();
@@ -139,7 +141,7 @@ public:
 		}
 		else {
 			// process positive number
-			_scale = int(sw::universal::findMostSignificantBit(rhs)) - 1;
+			_scale = int(sw::universal::find_msb(rhs)) - 1;
 			uint64_t _fraction_without_hidden_bit = uint64_t(_scale == 0 ? 0 : (rhs << (64 - _scale)));
 			_fraction = copy_integer_fraction<fbits>(_fraction_without_hidden_bit);
 			_nrOfBits = fbits;
@@ -170,7 +172,7 @@ public:
 		}
 		else {
 			reset();
-			_scale = static_cast<int>(sw::universal::findMostSignificantBit(rhs)) - 1;
+			_scale = static_cast<int>(sw::universal::find_msb(rhs)) - 1;
 			uint64_t _fraction_without_hidden_bit = _scale == 0 ? 0ull : (rhs << (64 - _scale)); // the scale == -1 case is handled above
 			_fraction = copy_integer_fraction<fbits>(_fraction_without_hidden_bit);
 			_nrOfBits = fbits;
