+++ include/universal/native/ieee754_numeric.hpp
@@ -1,0 +1,62 @@
+#pragma once
+// ieee754_numeric.hpp: numeric helper functions for IEEE-754 native types
+//
+// Copyright (C) 2017-2023 Stillwater Supercomputing, Inc.
+//
+// This file is part of the universal numbers project, which is released under an MIT Open Source license.
+
+namespace sw { namespace universal {
+
+	////////////////////////////////////////////////////////////////////////
+	// numerical helpers
+
+	// return the Unit in the Last Position
+	template<typename Real,
+		typename = typename ::std::enable_if< ::std::is_floating_point<Real>::value, Real >::type
+	>
+	inline Real ulp(const Real& a) {
+		return std::nextafter(a, a + a/2.0f) - a;
+	}
+
+	// check if the floating-point number is zero
+	template<typename Real,
+		typename = typename ::std::enable_if< ::std::is_floating_point<Real>::value, Real >::type
+	>
+	inline bool iszero(const Real& a) {
+		return (std::fpclassify(a) == FP_ZERO);
+	}
+
+	// compile time power of 2
+	template<typename Real, size_t powerOfTwo,
+		typename = typename ::std::enable_if< ::std::is_floating_point<Real>::value, Real >::type
+	>
+	inline constexpr Real ipow() {
+		Real base = 2.0f;
+		Real result = 1.0f;
+		size_t exp = powerOfTwo;
+		for (;;) {
+			if (exp & 1) result *= base;
+			exp >>= 1;
+			if (!exp) break;
+			base *= base;
+		}
+		return result;
+	}
+
+	// run time fast power of 2 with positive exponent
+	template<typename Real,
+		typename = typename ::std::enable_if< ::std::is_floating_point<Real>::value, Real >::type
+	>
+	inline constexpr Real ipow(size_t exp) {
+		Real base = 2.0f;
+		Real result = 1.0f;
+		for (;;) {
+			if (exp & 1) result *= base;
+			exp >>= 1;
+			if (!exp) break;
+			base *= base;
+		}
+		return result;
+	}
+
+}} // namespace sw::universal
