+++ include/universal/native/native.hpp
@@ -1,17 +1,16 @@
 #pragma once
 // native standard header aggregating all native type introspection and manipulation functionality
 //
-// Copyright (C) 2017-2021 Stillwater Supercomputing, Inc.
+// Copyright (C) 2017-2023 Stillwater Supercomputing, Inc.
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
 #ifndef _NATIVE_
 #define _NATIVE_
 
+#include <universal/utility/find_msb.hpp>
+#include <universal/utility/boolean_logic_operators.hpp>
 #include <universal/native/integers.hpp>
 #include <universal/native/ieee754.hpp>
-#include <universal/native/manipulators.hpp>
-#include <universal/native/bit_functions.hpp>
-#include <universal/native/boolean_logic_operators.hpp>
 #include <universal/native/subnormal.hpp>
 
 #endif
