+++ include/universal/number/float/quire.hpp
@@ -1,10 +1,10 @@
 #pragma once
 // quire.hpp: definition of a parameterized quire configurations for IEEE float/double/long double
 //
-// Copyright (C) 2017-2022 Stillwater Supercomputing, Inc.
+// Copyright (C) 2017-2023 Stillwater Supercomputing, Inc.
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
-#include <universal/native/boolean_logic_operators.hpp>
+#include <universal/utility/boolean_logic_operators.hpp>
 #include <universal/internal/value/value.hpp>
 #include <universal/number/quire/exceptions.hpp>
 
@@ -104,7 +104,7 @@ public:
 		_sign = rhs & 0x8000000000000000;
 		unsigned long long magnitude;
 		magnitude = _sign ? -rhs : rhs;
-		unsigned msb = findMostSignificantBit(magnitude);
+		unsigned msb = find_msb(magnitude);
 		if (msb > half_range + capacity) {
 			throw operand_too_large_for_quire{};
 		}
@@ -127,7 +127,7 @@ public:
 	}
 	quire& operator=(long long unsigned rhs) {
 		reset();
-		unsigned msb = findMostSignificantBit(rhs);
+		unsigned msb = find_msb(rhs);
 		if (msb > half_range + capacity) {
 			throw operand_too_large_for_quire{};
 		}
