+++ include/universal/utility/boolean_logic_operators.hpp
@@ -1,7 +1,7 @@
 #pragma once
 //  boolean_logic_operators.hpp : full set of boolean logic operators
 //
-// Copyright (C) 2017-2021 Stillwater Supercomputing, Inc.
+// Copyright (C) 2017-2023 Stillwater Supercomputing, Inc.
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
