+++ include/universal/utility/find_msb.hpp
@@ -1,7 +1,7 @@
 #pragma once
-// bit_functions.hpp: definitions of helper functions for bit operations on integers
+// find_msb.hpp: helper functions to find the most significant bit of an integer
 //
-// Copyright (C) 2017-2021 Stillwater Supercomputing, Inc.
+// Copyright (C) 2017-2023 Stillwater Supercomputing, Inc.
 //
 // This file is part of the universal numbers project, which is released under an MIT Open Source license.
 #include <cstdint>
@@ -21,7 +21,7 @@ namespace sw { namespace universal {
 /// </summary>
 /// <param name="x">value to</param>
 /// <returns> position of MSB that is set. LSB is defined to be at position 1, so no bits set returns 0</returns>
-inline constexpr unsigned int findMostSignificantBit(unsigned int x) {
+inline constexpr unsigned int find_msb(unsigned int x) {
 	// find the first non-zero bit
 	unsigned int base = 0;
 	if (x & 0xFFFF0000) { base += 16; x >>= 16; }
@@ -40,7 +40,7 @@ inline constexpr unsigned int findMostSignificantBit(unsigned int x) {
 /// </summary>
 /// <param name="x">value to</param>
 /// <returns> position of MSB that is set. LSB is defined to be at position 1, so no bits set returns 0</returns>
-inline constexpr unsigned int findMostSignificantBit(unsigned long x) {
+inline constexpr unsigned int find_msb(unsigned long x) {
 	// find the first non-zero bit
 	unsigned int base = 0;
 	if (x & 0xFFFF0000) { base += 16; x >>= 16; }
@@ -54,7 +54,7 @@ inline constexpr unsigned int findMostSignificantBit(unsigned long x) {
 	return base + bval;
 }
 
-inline constexpr unsigned int findMostSignificantBit(unsigned long long x) {
+inline constexpr unsigned int find_msb(unsigned long long x) {
 	// find the first non-zero bit
 	unsigned int base = 0;
 	if (x & 0xFFFFFFFF00000000) { base += 32; x >>= 32; }
@@ -72,7 +72,7 @@ inline constexpr unsigned int findMostSignificantBit(unsigned long long x) {
 //////////////////////////// SIGNED integer types ////////////////////////
 
 
-inline constexpr unsigned int findMostSignificantBit(signed char x) {
+inline constexpr unsigned int find_msb(signed char x) {
 	// find the first non-zero bit
 	uint8_t tmp = uint8_t(x);
 	unsigned int base = 0;
@@ -85,7 +85,7 @@ inline constexpr unsigned int findMostSignificantBit(signed char x) {
 	return base + bval;
 }
 
-inline constexpr unsigned int findMostSignificantBit(short x) {
+inline constexpr unsigned int find_msb(short x) {
 	// find the first non-zero bit
 	uint16_t tmp = uint16_t(x);
 	unsigned int base = 0;
@@ -99,7 +99,7 @@ inline constexpr unsigned int findMostSignificantBit(short x) {
 	return base + bval;
 }
 
-inline constexpr unsigned int findMostSignificantBit(int x) {
+inline constexpr unsigned int find_msb(int x) {
 	// find the first non-zero bit
 	uint32_t tmp = uint32_t(x);
 	unsigned int base = 0;
@@ -114,7 +114,7 @@ inline constexpr unsigned int findMostSignificantBit(int x) {
 	return base + bval;
 }
 
-inline constexpr unsigned int findMostSignificantBit(long x) {
+inline constexpr unsigned int find_msb(long x) {
 	// find the first non-zero bit
 	uint32_t tmp = uint32_t(x);
 	unsigned int base = 0;
@@ -130,7 +130,7 @@ inline constexpr unsigned int findMostSignificantBit(long x) {
 	return base + bval;
 }
 
-inline constexpr unsigned int findMostSignificantBit(long long x) {
+inline constexpr unsigned int find_msb(long long x) {
 	// find the first non-zero bit
 	uint64_t tmp = uint64_t(x);
 	unsigned int base = 0;
